+++ Source/Models/Payloads/Dragon.h
@@ -9,8 +9,8 @@ namespace Marsy::Models::Payloads
     struct Dragon
     {
         std::optional<std::string> capsule = std::nullopt;
-        std::optional<float> massReturnedKilograms = std::nullopt;
-        std::optional<float> massReturnedPounds = std::nullopt;
+        std::optional<double> massReturnedKilograms = std::nullopt;
+        std::optional<double> massReturnedPounds = std::nullopt;
         std::optional<int> flightTimeSeconds = std::nullopt;
         std::optional<std::string> manifest = std::nullopt;
         std::optional<bool> waterLanding = std::nullopt;
